pyproject.toml
[project]
name = "od-training"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Mizuki", email = "mbaba@kanolab.net" }
]
dependencies = [
    "pandas>=2.0.3",
    "ja_sentence_segmenter>=0.0.2",
    "torch>=2.0.1",
    "transformers>=4.32.1",
    "mlflow>=2.6.0",
    "pytorch-lightning>=2.0.8",
    "fugashi>=1.3.0",
    "ipadic>=1.0.0",
    "tensorboard>=2.14.0",
    "pyyaml>=6.0.1",
    "scikit-learn>=1.3.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
